@@ -1,7 +1,7 @@
 {
-    "name": "laravel/laravel",
+    "name": "kolab/kolab4",
     "type": "project",
-    "description": "The Laravel Framework.",
+    "description": "Kolab 4",
     "keywords": [
         "framework",
         "laravel"
@@ -14,37 +14,32 @@
         }
     ],
     "require": {
-        "php": "^7.3",
-        "barryvdh/laravel-dompdf": "^0.8.6",
-        "doctrine/dbal": "^2.13",
-        "dyrynda/laravel-nullable-fields": "*",
-        "fideloper/proxy": "^4.0",
-        "guzzlehttp/guzzle": "^7.3",
+        "php": "^8.0",
+        "barryvdh/laravel-dompdf": "^1.0.0",
+        "doctrine/dbal": "^3.3.2",
+        "dyrynda/laravel-nullable-fields": "^4.2.0",
+        "guzzlehttp/guzzle": "^7.4.1",
         "kolab/net_ldap3": "dev-master",
-        "laravel/framework": "6.*",
-        "laravel/horizon": "^3",
-        "laravel/passport": "^9",
-        "laravel/tinker": "^2.4",
-        "mlocati/spf-lib": "^3.0",
-        "mollie/laravel-mollie": "^2.9",
+        "laravel/framework": "^9.0",
+        "laravel/horizon": "^5.9",
+        "laravel/passport": "^10.3",
+        "laravel/tinker": "^2.7",
+        "mlocati/spf-lib": "^3.1",
+        "mollie/laravel-mollie": "^2.19",
         "moontoast/math": "^1.2",
-        "morrislaptop/laravel-queue-clear": "^1.2",
         "pear/crypt_gpg": "^1.6.6",
-        "silviolleite/laravelpwa": "^2.0",
-        "spatie/laravel-translatable": "^4.2",
+        "predis/predis": "^1.1.10",
+        "spatie/laravel-translatable": "^5.2",
         "spomky-labs/otphp": "~4.0.0",
-        "stripe/stripe-php": "^7.29",
-        "swooletw/laravel-swoole": "^2.6"
+        "stripe/stripe-php": "^7.29"
     },
     "require-dev": {
-        "beyondcode/laravel-er-diagram-generator": "^1.3",
-        "code-lts/doctum": "^5.1",
-        "kirschbaum-development/mail-intercept": "^0.2.4",
-        "laravel/dusk": "~6.15.0",
-        "nunomaduro/larastan": "^0.7",
-        "phpstan/phpstan": "^0.12",
+        "code-lts/doctum": "^5.5.1",
+        "laravel/dusk": "~6.22.0",
+        "nunomaduro/larastan": "^2.0",
+        "phpstan/phpstan": "^1.4",
         "phpunit/phpunit": "^9",
-        "squizlabs/php_codesniffer": "3.*"
+        "squizlabs/php_codesniffer": "^3.6"
     },
     "config": {
         "optimize-autoloader": true,
