@@ -9,47 +9,28 @@ use Illuminate\Support\Facades\DB;
 
 class Handler extends ExceptionHandler
 {
-    /**
-     * A list of the exception types that are not reported.
-     *
-     * @var array
-     */
+    /** @var string[] A list of the exception types that are not reported */
     protected $dontReport = [
         \Laravel\Passport\Exceptions\OAuthServerException::class,
         \League\OAuth2\Server\Exception\OAuthServerException::class
     ];
 
-    /**
-     * A list of the inputs that are never flashed for validation exceptions.
-     *
-     * @var array
-     */
+    /** @var string[] A list of the inputs that are never flashed for validation exceptions */
     protected $dontFlash = [
         'password',
         'password_confirmation',
     ];
 
-    /**
-     * Report or log an exception.
-     *
-     * @param \Exception $exception
-     *
-     * @return void
-     */
-    public function report(Exception $exception)
-    {
-        parent::report($exception);
-    }
 
     /**
      * Render an exception into an HTTP response.
      *
      * @param \Illuminate\Http\Request $request
-     * @param \Exception               $exception
+     * @param \Throwable               $exception
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function render($request, Exception $exception)
+    public function render($request, \Throwable $exception)
     {
         // Rollback uncommitted transactions
         while (DB::transactionLevel() > 0) {
