@@ -14,9 +14,8 @@ use Illuminate\Database\Eloquent\Model;
  */
 class WalletSetting extends Model
 {
-    protected $fillable = [
-        'wallet_id', 'key', 'value'
-    ];
+    /** @var string[] The attributes that are mass assignable */
+    protected $fillable = ['wallet_id', 'key', 'value'];
 
     /**
      * The wallet to which this setting belongs.
