@@ -5,13 +5,10 @@ namespace Tests\Unit\Mail;
 use App\Mail\PaymentMandateDisabled;
 use App\Wallet;
 use App\User;
-use Tests\MailInterceptTrait;
 use Tests\TestCase;
 
 class PaymentMandateDisabledTest extends TestCase
 {
-    use MailInterceptTrait;
-
     /**
      * Test email content
      */
@@ -22,7 +19,7 @@ class PaymentMandateDisabledTest extends TestCase
 
         \config(['app.support_url' => 'https://kolab.org/support']);
 
-        $mail = $this->fakeMail(new PaymentMandateDisabled($wallet, $user));
+        $mail = $this->renderMail(new PaymentMandateDisabled($wallet, $user));
 
         $html = $mail['html'];
         $plain = $mail['plain'];
@@ -33,7 +30,7 @@ class PaymentMandateDisabledTest extends TestCase
         $supportLink = sprintf('<a href="%s">%s</a>', $supportUrl, $supportUrl);
         $appName = \config('app.name');
 
-        $this->assertMailSubject("$appName Auto-payment Problem", $mail['message']);
+        $this->assertSame("$appName Auto-payment Problem", $mail['subject']);
 
         $this->assertStringStartsWith('<!DOCTYPE html>', $html);
         $this->assertTrue(strpos($html, $user->name(true)) > 0);
